feat: add isActive props and refactor
